### 2018-08-24 #914 @whedon
[PRE REVIEW]: DataDepsGenerators.jl: making reusing data easy by automatically generating DataDeps.jl registration code
**Submitting author:** @oxinabox (<a href="http://orcid.org/0000-0003-1386-1646">Lyndon White</a>)
**Repository:** <a href="https://github.com/oxinabox/DataDepsGenerators.jl/" target ="_blank">https://github.com/oxinabox/DataDepsGenerators.jl/</a>
**Version:** v0.4.0
