### 2018-08-27 91aa0c8
See if this improves affiliation typesetting
