### 2018-08-30 49ee780
Documentation in README (#49)
* Documentation in README

* Minor details

* Modify readme

* Modify readme 1
