+++ test/JSONLD/JSONLD.jl
@@ -6,7 +6,7 @@ using ReferenceTests
 @testset "JSON-LD Web" begin
     @test_reference "../references/JSONLD_Web Kaggle.txt" generate(JSONLD_Web(), "https://www.kaggle.com/stackoverflow/stack-overflow-2018-developer-survey")
     @test_reference "../references/JSONLD_Web Zenodo.txt" generate(JSONLD_Web(), "https://zenodo.org/record/1287281")
-    @test_reference "../references/JSONLD_Web ICRISAT.txt" generate(JSONLD_Web(), "http://dataverse.icrisat.org/dataset.xhtml?persistentId=doi:10.21421/D2/ZS6XX1")
+    @test_reference "../references/JSONLD_Web Dataverse.txt" generate(JSONLD_Web(), "https://dataverse.harvard.edu/dataset.xhtml?persistentId=hdl:1902.1/15580")
     
     @test_reference "../references/JSONLD_Web Figshare.txt" generate(JSONLD_Web(), "https://figshare.com/articles/_shows_examples_of_coordinated_and_uncoordinated_motion_for_dangerous_and_non_dangerous_crowd_behavior_/186003")
 end
