### 2018-08-31 #914 @juliohm
@oxinabox I passed my defense! \o/ Everything is super busy around here though as I am relocating, moving out from California, and submitting the final draft of the dissertation to the university. I thought I would have time to review your paper, but I am super busy wrapping up. Will join a vacation trip in the next weeks, which will literally put me offline for some time.

I appreciate if other Julians from your list of suggested reviewers can help: @vchuravy @ahwillia @ChrisRackauckas @ninjin
