### 2018-08-31 ab1cfdd
Update paper.md
