### 2018-09-04 a81c73b
Quieten Integration tests
