+++ test/CKAN.jl
@@ -7,7 +7,7 @@ using ReferenceTests
 @testset "CKAN Demo test" begin
     registration_code = generate(CKAN(), "https://demo.ckan.org/dataset/gold-prices")
 
-    @testset "Integration Test" begin
+    @suppress @testset "Integration Test" begin
         eval(parse(registration_code))
         @test length(collect(readdir(datadep"Gold Prices in London 1950-2008 (Monthly)"))) > 0
     end
