+++ test/CKAN.jl
@@ -1,7 +1,7 @@
 using DataDeps
 using DataDepsGenerators
 using Base.Test
-
+using Suppressor
 using ReferenceTests
 
 @testset "CKAN Demo test" begin
