+++ test/UCI.jl
@@ -1,13 +1,13 @@
 using DataDeps
 using DataDepsGenerators
 using Base.Test
-
+usinng Suppressor
 using ReferenceTests
 
 @testset "UCI Air Quality" begin
     registration_code = generate(UCI(), "Air+Quality")
 
-    @testset "Integration Test" begin
+    @suppress @testset "Integration Test" begin
         eval(parse(registration_code))
         @test length(collect(readdir(datadep"Air Quality Data Set (UCI ML Repository)"))) > 0
     end
