+++ test/CKAN.jl
@@ -1,6 +1,6 @@
+using Base.Test
 using DataDeps
 using DataDepsGenerators
-using Base.Test
 using Suppressor
 using ReferenceTests
 
