### 2018-09-14 9f1a654
=drop 0.6 support
