+++ .travis.yml
@@ -3,14 +3,15 @@ language: julia
 os:
   - linux
 julia:
-  - 0.6
+  - 0.7
+  - 1.0
   - nightly
 notifications:
   email: false
 git:
   depth: 99999999
 env:
-  - DATADEPS_ALWAY_ACCEPT=true
+  - DATADEPS_ALWAYS_ACCEPT=true
 
 ## uncomment the following lines to allow failures on nightly julia
 ## (tests will run but not make your overall status red)
