+++ appveyor.yml
@@ -11,7 +11,7 @@ platform:
 
 # # Uncomment the following lines to allow failures on nightly julia
 # # (tests will run but not make your overall status red)
-# matrix:
+ matrix:
    allow_failures:
    - julia_version: nightly
 
