+++ src/DataOneV2/DataOneV2.jl
@@ -12,7 +12,7 @@ published_date(repo::DataOneV2, mainpage) = try pub_date(repo, mainpage) catch m
 
 function license(repo::DataOneV2, mainpage)
     try
-        license_ele = matchall(sel"intellectualRights", mainpage.root)
+        license_ele = eachmatch(sel"intellectualRights", mainpage.root)
         if length(license_ele) > 0
             text_only(first(license_ele))
         else
@@ -25,7 +25,7 @@ end
 
 function paper_title_(repo::DataOneV2, mainpage)
     try
-        paper_title_ele = matchall(sel"title", mainpage.root)
+        paper_title_ele = eachmatch(sel"title", mainpage.root)
         if length(paper_title_ele) > 0
             text_only(first(paper_title_ele))
         else
