+++ src/JSONLD/JSONLD_Web.jl
@@ -3,7 +3,7 @@ struct JSONLD_Web <: JSONLD end
 function mainpage_url(repo::JSONLD_Web, dataname)
     page=getpage(dataname)
     pattern = sel"script[type=\"application/ld+json\"]"
-    jsonld_blocks = matchall(pattern, page.root)
+    jsonld_blocks = eachmatch(pattern, page.root)
     if length(jsonld_blocks)==0
         throw(GeneratorError(repo, "No JSON-LD Linked Data Found"))
     end
