+++ src/DataDepsGenerators.jl
@@ -78,15 +78,15 @@ get_urls(::DataRepo, mainpage) = missing
 
 include("utils.jl")
 include("generic_extractors.jl")
-include("UCI.jl")
-include("GitHub.jl")
-include("DataDryad.jl")
-include("DataOneV1.jl")
-include("DataOneV2/DataOneV2.jl")
-include("CKAN.jl")
-include("DataCite.jl")
-include("Figshare.jl")
-include("JSONLD/JSONLD.jl")
+include("APIs/UCI.jl")
+include("APIs/GitHub.jl")
+include("APIs/DataDryad.jl")
+include("APIs/DataOneV1.jl")
+include("APIs/DataOneV2/DataOneV2.jl")
+include("APIs/CKAN.jl")
+include("APIs/DataCite.jl")
+include("APIs/Figshare.jl")
+include("APIs/JSONLD/JSONLD.jl")
 
 function aggregate(metadatas)
 
