+++ src/DataDepsGenerators.jl
@@ -8,7 +8,8 @@ using Dates
 using InteractiveUtils: subtypes
 
 export generate, citation_text, remove_cite_version
-export UCI, GitHub, DataDryad, DataOneV1, DataOneV2, CKAN, DataCite, Figshare, JSONLD
+export UCI, GitHub, DataDryad, DataOneV1, DataOneV2, CKAN, DataCite, Figshare, JSONLD_Web, JSONLD_DOI
+
 
 
 
