### 2018-09-14 0d5d47d
=tidy up generate with doc strings
