+++ test/combine.jl
@@ -13,4 +13,8 @@ using ReferenceTests
     @test_reference "references/Combining Stack Overflow.txt" generate("https://www.kaggle.com/stackoverflow/stack-overflow-2018-developer-survey")
     
     @test_reference "references/Combining Chile Activism.txt" generate("https://figshare.com/articles/Youth_Activism_in_Chile_from_urban_educational_inequalities_to_experiences_of_living_together_and_solidarity/6504206")
+
+    @test_reference "references/Combining oisst.txt" generate("https://data.nodc.noaa.gov/cgi-bin/iso?id=gov.noaa.ncdc:C00844","oisst")
+
+
 end
