+++ test/UCI.jl
@@ -17,7 +17,7 @@ end
 
 @test_reference "references/UCI auto mpg.txt" generate(UCI(), "https://archive.ics.uci.edu/ml/datasets/auto+mpg")
 @test_reference "references/UCI banking marketting.txt" generate(UCI(), "bank+marketing")
-@test_reference "references/UCI BHP.txt" generate(UCI(), "Burst+Header+Packet+%28BHP%29+flooding+attack+on+Optical+Burst+Switching+%28OBS%29+Network", shortname="Burst Header Packet (UCI)")
+@test_reference "references/UCI BHP.txt" generate(UCI(), "Burst+Header+Packet+%28BHP%29+flooding+attack+on+Optical+Burst+Switching+%28OBS%29+Network", "Burst Header Packet (UCI)")
 
 @test_reference "references/UCI Adult.txt" generate(UCI(), "Adult")
 
