### 2018-09-14 3ac0f30
=simplify single case
