+++ Project.toml
@@ -9,13 +9,18 @@ version = "0.5.0"
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 Cascadia = "54eefc05-d75b-58de-a785-1a3403f0919f"
-DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Gumbo = "708ec375-b3d6-5a57-a7ce-8257bf98657a"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
-MD5 = "6ac74813-4b46-53a4-afec-0b5dc9d7885c"
-ReferenceTests = "324d217c-45ce-50fc-942e-d289b448e8cf"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
+
+[extras]
+MD5 = "6ac74813-4b46-53a4-afec-0b5dc9d7885c"
+DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
+ReferenceTests = "324d217c-45ce-50fc-942e-d289b448e8cf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["MD5", "DataDeps", "ReferenceTests", "Test"]
