### 2018-09-15 d8ac788
Fix deprecations
