### 2018-09-15 f044ee6
=switch to HTTP for page download to enable async
