+++ src/APIs/CKAN.jl
@@ -25,12 +25,12 @@ end
 data_fullname(::CKAN, mainpage) = mainpage["title"]
 
 function website(repo::CKAN, mainpage_url, mainpage)
-    replace(mainpage_url, "/api/3/action/package_show?id=", "/dataset/")
+    replace(mainpage_url, "/api/3/action/package_show?id=" =>  "/dataset/")
 end
 
 function mainpage_url(repo::CKAN, dataname)
     if startswith(dataname, "http")
-        url = replace(dataname, "/dataset/", "/api/3/action/package_show?id=")
+        url = replace(dataname, "/dataset/" =>  "/api/3/action/package_show?id=")
         return JSON.parse(text_only(getpage(url).root))["result"], dataname
     else
         # error("Please use a valid url")
