+++ src/APIs/UCI.jl
@@ -10,11 +10,11 @@ base_url(::UCI) = "https://archive.ics.uci.edu/ml/datasets/"
 function to_cite(::UCI, mainpage)
     section = text_only(Gumbo.children(last(eachmatch(sel"p + p.normal", mainpage.root))))
 
-    section = replace(section, "Available at: \n[Web Link]", "")
-    section = replace(section, r"\(?\s*\[Web Link\]\s*\)?","")
+    section = replace(section, "Available at: \n[Web Link]" => "")
+    section = replace(section, r"\(?\s*\[Web Link\]\s*\)?" => "")
     section *= "\nLichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science."
 
-    replace(section, r"\n+","\n")
+    replace(section, r"\n+" => "\n")
 end
 
 function description(::UCI, mainpage)
