+++ src/APIs/UCI.jl
@@ -24,7 +24,7 @@ end
 
 function data_fullname(::UCI, mainpage)
     data_fullname = text_only(eachmatch(sel".heading b", mainpage.root))
-    data_fullname*= " (UCI ML Repository)"
+    data_fullname *= " (UCI ML Repository)"
 end
 
 function get_urls(::UCI, mainpage)
