### 2018-09-20 0b1d1aa
=remove nondetermanistic test
