### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @ninjin
Undeclared dependency for Pkg3 on 0.7
On 0.7 and installed using Pkg3, `using DataDepsGenerators` fails with `ERROR: LoadError: ArgumentError: Package DataDepsGenerators does not have Missings in its dependencies`. This is currently blocking [JOSS review](https://github.com/openjournals/joss-reviews/issues/921).
