### 2018-09-26 #921 @ninjin
Apologies for the delays, here is the current state of my review.

Formatting:

1. The citations look a bit odd to me, should “((Bezanson et al. 2014))” not be “(Bezanson et al. 2014)”?
2. A period is missing at the end of the third bullet in the first bullet list.
3. In the second to last paragraph there is what I think is a failed listing with hyphens following dots, this should probably be fixed.
4. I suspect that both of the cited works have DOIs assigned to them, Bezanson et al. (2014) I think was accepted by some IEEE journal. These DOIs should be added.

Content:

1. Are there not other similar packages for other programming languages? There is currently a bit of a lack of a wider context of open science, etc. and I think one or two sentences with a few cites would be very helpful for the readers.

Other than the above comments, the submitted paper reads very well.

Documentation:

1. No clear statement of need.
2. Misspellings: “consitantly”, “webbrowser”, “targetted”, etc. run a spell checker.
3. Awkward grammar in a number of places, just give it another read through.
4. While this is a small package, having at least a rudimentary `CONTRIBUTING.md` would be a good idea.

Functionality:

Currently blocked by [#62](https://github.com/oxinabox/DataDepsGenerators.jl/issues/62).
