### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @oxinabox
The short when I sumbitted to JOSS I was looking at tagged release for julia 0.6.
which was DataDepsGenerators.jl v0.4.0

I believe master works with julia 0.7,
but it isn't extensively tested,
as julia 0.7 almost-working was merged because I was having a fight with @femtocleaner[bot].
#60 finishes it off.

Once #60 is done I will be more confidant that it will fully support julia 0.7 and julia 1.0.
#60 works locally for me, but is still having some travis issues.
