### 2018-09-26 #921 @oxinabox
Thanks, I will attend to these in the next couple of days.
