### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @ghost
My bad, do you want me to proceed with an 0.6 review of 0.4.0?
