### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @oxinabox
Up to you.
Feel free to wait for #60 to be merged.
It will be a smoother experience.
Because async works correctly.

And all the tests will be updated to match the fact that some upstream repos have taken files down etc.
