### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @ghost
Thanks, will proceed with 0.7 then as I expect that is where most of your users will be now that 0.6 is actively discouraged.
