+++ test/Figshare.jl
@@ -10,8 +10,8 @@ using ReferenceTests
     @testset "Chile Activism" begin
         registration_code = generate(Figshare(), "https://figshare.com/articles/Youth_Activism_in_Chile_from_urban_educational_inequalities_to_experiences_of_living_together_and_solidarity/6504206")
 
-        @suppress @testset "Integration Test" begin
-            eval(parse(registration_code))
+        @testset "Integration Test" begin
+            eval(Meta.parse(registration_code))
             @test length(collect(readdir(datadep"Youth Activism in Chile from urban educational inequalities to experiences of living together and solidarity"))) > 0
         end
         
