+++ test/GitHub.jl
@@ -1,9 +1,6 @@
 using DataDeps
 using DataDepsGenerators
 using Test
-
-using Suppressor
-
 using ReferenceTests
 
 """
@@ -18,8 +15,8 @@ end
     @testset "538 March" begin
         registration_code = generate(GitHub(), "https://github.com/fivethirtyeight/data/tree/master/march-madness-predictions-2015")
 
-        @suppress @testset "Integration Test" begin
-            eval(parse(registration_code))
+        @testset "Integration Test" begin
+            eval(Meta.parse(registration_code))
             @test length(collect(readdir(datadep"March Madness Predictions"))) > 0
         end
 
@@ -39,8 +36,8 @@ end
     @testset "Pres Camp" begin
         registration_code = generate(GitHub(), "https://github.com/BuzzFeedNews/presidential-campaign-contributions")
         
-		@suppress @testset "Integration Test" begin
-			eval(parse(registration_code)) # evaluate the new code
+		@testset "Integration Test" begin
+			eval(Meta.parse(registration_code)) # evaluate the new code
 			@test length(collect(readdir(datadep"Presidential Campaign Contributions"))) > 0
 		end
 		
