### 2018-09-26 #921 @oxinabox
> I suspect that both of the cited works have DOIs assigned to them, Bezanson et al. (2014) I think was accepted by some IEEE journal. These DOIs should be added.

the DataDeps work is still under-review (and is taking a truly long time) and does not yet have a DOI.
I have searched [IEEE](https://ieeexplore.ieee.org/search/searchresult.jsp?newsearch=true&queryText=Bezanson)
and I can not find 2014 julia paper there.
But it appears now the 2017 SIAM paper [is now the preferred citation for Julia, so I will update it to that](https://julialang.org/publications/),
that one does have a DOI.
