### 2018-09-26 #921 @ninjin
Your are indeed correct, it was a SIAM paper and not IEEE.
