### 2018-09-26 67d6efe
remove hypothosis tests dep
