### 2018-09-26 198e557
Julia paper should be SIAM version
