+++ paper/paper.bib
@@ -1,15 +1,14 @@
-
-@Article{Julia,
-  Title                    = {{J}ulia: A Fresh Approach to Numerical Computing},
-  Author                   = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
-  Year                     = {2014},
-  Eprint                   = {1411.1607},
-  Eprintclass              = {cs.MS},
-  Eprinttype               = {arXiv},
-  Keywords                 = {tools},
-  Owner                    = {20361362},
-  Timestamp                = {2015.12.18},
-  Url                      = {http://arxiv.org/abs/1411.1607}
+@article{Julia,
+author = {Bezanson, J. and Edelman, A. and Karpinski, S. and Shah, V.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+URL = {https://doi.org/10.1137/141000671},
+eprint = {https://doi.org/10.1137/141000671}
 }
 
 
