### 2018-09-26 oxinabox/DataDepsGenerators.jl#62 @oxinabox
Closed in #60.

Julia 0.7 compat version is waiting to merge
https://github.com/JuliaLang/METADATA.jl/pull/18409
