### 2018-09-26 808437b
Update paper.md
