### 2018-09-26 7268aee
Update paper.md
