+++ paper/paper.md
@@ -107,7 +107,7 @@ Examples include:
 
 To the best of our knowledge at present their is not any unifying R package that supports anywhere near the range of data repostories supported by DataDepsGenerators.jl.
 Contemporaneously, with the creation of DataDepsGenerator.jl,
-the was proposed package to acquire data based on a DOI [doidata](https://github.com/ropenscilabs/doidata).
+the was proposed package ([doidata](https://github.com/ropenscilabs/doidata)) to acquire data based on a DOI.
 While this has yet to eventuate into usable software, several of the discussions relating to it were insightful,
 and contributed to the functionality of DataDepsGenerators.jl
 
