### 2018-09-27 c4e3511
Update paper.bib
