+++ paper/paper.bib
@@ -11,6 +11,19 @@ URL = {https://doi.org/10.1137/141000671},
 eprint = {https://doi.org/10.1137/141000671}
 }
 
+@article{suppdata,
+  doi = {10.21105/joss.00721},
+  url = {https://doi.org/10.21105/joss.00721},
+  year  = {2018},
+  month = {may},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {25},
+  pages = {721},
+  author = {William D Pearse and Scott A Chamberlain},
+  title = {Suppdata: Downloading Supplementary Data from Published Manuscripts},
+  journal = {Journal of Open Source Software}
+}
 
 @Article{2018arXiv180801091W,
   author        = {{White}, L. and {Togneri}, R. and {Liu}, W. and {Bennamoun}, M.},
