### 2018-09-27 bbd3fc6
Update paper.md
