### 2018-09-27 f3c130c
Create CONTRIBUTING.md
