+++ CONTRIBUTING.md
@@ -1,0 +1,38 @@
+# Contributing to DataDepsGenerators.jl
+
+To emphasise this at the start:
+**if you have an open PR or Issue, feel free to bump it, or ask for a status update**
+every few days if you've not gotten a response yet,
+or even if it has been been a while.
+
+There is a good chance I have lost track of your issue/PR in the GitHub notifications screen;
+and a reminder would be appreciated.
+Don't worry that you might annoy me, I will let you know if you are;
+or more likely I will tell you when it is a good time to bump me again (e.g. next week).
+
+
+## Issues / Bug reports
+
+Issues should be raised on this GitHub at:
+https://github.com/oxinabox/DataDepsGenerators.jl/issues
+
+It is worth knowing that some bugs will need help from the data repositories support team.
+
+
+## Questions
+
+If you have any questions about DataDepsGenerators.jl,
+feel encouraged to ping @oxinabox on the [JuliaLang Slack](https://slackinvite.julialang.org/) in the `#data` channel,
+or on the [JuliaLang Discourse forum](https://discourse.julialang.org/c/domain/data) again in the data channel.
+
+It is also permissible to ask questions by opening issues on this GitHub Repo.
+Ideally, such questions be phrased as requests for documentation.
+If your question is not answered by the docs(/readme), then it is a sign the docs could be improved.
+
+## Contributing code
+Code contributions are welcome and encouraged.
+Especially adding support for a new data repository.
+
+Code contributions should be made in the normal way of making a pull request.
+In general they should try to match the style of the code already present.
+This follows the [julialang/julia repo conventions](https://github.com/JuliaLang/julia/blob/master/CONTRIBUTING.md#code-formatting-guidelines).
