### 2018-09-27 818b42a
Update CONTRIBUTING.md
