+++ CONTRIBUTING.md
@@ -17,19 +17,20 @@ Issues should be raised on this GitHub at:
 https://github.com/oxinabox/DataDepsGenerators.jl/issues
 
 It is worth knowing that some bugs will need help from the data repositories support team.
+I can advise on that; often I'll want to take it up with them.
 
 
 ## Questions
 
 If you have any questions about DataDepsGenerators.jl,
-feel encouraged to ping @oxinabox on the [JuliaLang Slack](https://slackinvite.julialang.org/) in the `#data` channel,
+feel encouraged to ping **@oxinabox** on the [JuliaLang Slack](https://slackinvite.julialang.org/) in the `#data` channel,
 or on the [JuliaLang Discourse forum](https://discourse.julialang.org/c/domain/data) again in the data channel.
 
 It is also permissible to ask questions by opening issues on this GitHub Repo.
 Ideally, such questions be phrased as requests for documentation.
 If your question is not answered by the docs(/readme), then it is a sign the docs could be improved.
 
-## Contributing code
+## Contributing Code
 Code contributions are welcome and encouraged.
 Especially adding support for a new data repository.
 
