+++ README.md
@@ -4,16 +4,29 @@ AppVeyor Master: [![Build status](https://ci.appveyor.com/api/projects/status/2q
 
 
 **Generating registration blocks for [DataDeps.jl](https://github.com/oxinabox/DataDeps.jl) in one key press.**
+An example of use [is in this blog-post](http://white.ucc.asn.au/2018/01/18/DataDeps.jl-Repeatabled-Data-Setup-for-Repeatable-Science.html#example-3-538-avenegers-comic-book-characters--datadepsgeneratorsjl)
 
-This package should not be used as a dependancy
-Instead the interactive features of this package should be used from the Julia REPL,
-to get a good registration block, which is output to a file (or STDOUT),
+DataDepsGenerators.jl is a tool written to help users of the Julia programming language
+to observe best practices when making use of published datasets.
+Using the metadata present in published datasets, it generates the code for the data dependency registration blocks required by DataDeps.jl [@2018arXiv180801091W].
+These registration blocks are effectively executable metadata,
+which can be resolved by DataDeps.jl to download the dataset.
+They include a message that is displayed to the user whenever the data set is automatically downloaded.
+This message should include provenance information on the dataset,
+so that downstream users know its original source and details on its processing.
+
+
+
+This package should not be used as a direct dependency.
+Instead its interactive features should be used from the Julia REPL,
+to generate a good registration code block
 which can be added to your package.
 
-While it can be used to directly create and invoke a registration,
-this use handy for interactive and prototyping use,
-but not great for packages, as it involves triggering a webscraper everytime the package is loaded.
-Not to mention DataDepsGenerators has a pretty heavy set of dependencies,
+The registration block can be immediately evaluated using `eval(Meta.parse(generate(...)))`
+which is handy for interactive prototyping,
+it is not great to put this code in a package packages;
+as it involves triggering web requests every time the package is loaded.
+Not to mention DataDepsGenerators.jl has a pretty heavy set of dependencies,
 that you don't really want weighing down your package.
 
 
@@ -21,50 +34,65 @@ that you don't really want weighing down your package.
 But it is up to you make sure it is right.
 The code it generates isn't always the cleanest.
 It may capture too much, or too little information.
-It might get things wrong (particularly when it is scraping websites that are not very consitantly formatted).
-After generating make sure to take a few minutes to check the code is code you are happy with.
+It might get things wrong (particularly when the metadata retrieved is wrong).
+You should take a few moments to check you are happy with the registration block code generated.
 Make a few tweaks, and it should be good to go.
 
-An example of use [is in this blog-post](http://white.ucc.asn.au/2018/01/18/DataDeps.jl-Repeatabled-Data-Setup-for-Repeatable-Science.html#example-3-538-avenegers-comic-book-characters--datadepsgeneratorsjl)
 
 ## Usage
 
 All usage revolves around the `generate` command. `generate()` is an overloaded method with two ways of usage:
 
-### Default Usage:
+### Normal Usage:
+The normal way to use the package is to use the the one or two argument form.
+
 ```julia
-generate("https://datadryad.org/resource/doi:10.5061/dryad.74699")
+generate(id_or_url, [datadep_name])::String
 ```
 
-- the first argument is the `id_or_url`
-     - this lets us know which dataset (from that repository) is being targetted.
[…]
