### 2018-09-27 #921 @oxinabox
For my own records:
**Formatting** issues are fixed.

**Content**  issues are fixed, at least for R I added a section.
I don't know much about this space for Python.

**Documentation:**
Will be fixed once https://github.com/oxinabox/DataDepsGenerators.jl/pull/64
is merged

> No clear statement of need.

I took the first paragraph from the paper and added it to the docs,
since I think that sums it up pretty well.

> Misspellings: “consitantly”, “webbrowser”, “targetted”, etc. run a spell checker.

Done

> Awkward grammar in a number of places, just give it another read through.

Done

> While this is a small package, having at least a rudimentary CONTRIBUTING.md would be a good idea

Done
--------------------------------------

Re Functionality being blocked.
That is now resolved for julia 0.7+/1.0  in the newly released DataDepsGenerators v0.5.0
The DataDepsGenerators  0.4.0 release can be used on 0.6, but I agree with  @ninjin  it is better for him to review the julia 1.0 compatible version since that will become the version most people will use.
