### 2018-10-02 b5ffca3
Update paper.md
