### 2018-10-22 df73c92
Syntax fixes for examples
Bonus, spelling “registation” → “registration”.
