### 2018-10-22 #921 @ninjin
I think GitHub is back online…

# Writing #

Nothing major, but some issues that should be addressed.

1. “leavages”
2. The CKAN entry has a bullet point split across two bullet points.
3. Same for JSON-LD.
4. “Julia” should always be capitalised.
5. “By generating code that the user may edit they may tweak the details as required.”, this one is a bit hard to parse, reformulate?
6. “If these APIs were to be access directly by a project, …”, not native, but “accessed”?
7. Maybe not use “things” in academic writing.
8. “the was proposed package (doidata) to acquire data based on a DOI.”, grammar.

# Code #

Pretty much works as intended, just some minor syntax issues in the README and some examples that were borked. Overall, I would say the code is good to go.

* [Syntax fix PR filed](https://github.com/oxinabox/DataDepsGenerators.jl/pull/66) (and merged)
* [Example issue](https://github.com/oxinabox/DataDepsGenerators.jl/issues/67)
