### 2018-10-23 oxinabox/DataDepsGenerators.jl#67 @oxinabox
Thanks. I will chase these up, unless @SebastinSanty  gets to it first
