### 2018-10-23 #921 @oxinabox
Fixed the paper. thank you for the feedback.
(At some point I misread the guidance on the capitalization of julia, getting it literally backwards,
and thought it was  **julia** for the language, and **Julia** for the compiler.
But indeed it is the other way round (which makes much more sense since lowercase matches the executable name).
