### 2018-10-24 ea462d3
Fix typos
