### 2018-10-24 42094aa
delete unused code
