### 2018-10-24 22c2272
Correct figshare example
