### 2018-10-24 87deb10
Don't run Integration on CKAN demo site
