### 2018-10-24 23f9d9c
Don't run integration tests on CKAN demo site
