### 2018-10-24 3038f01
remove extranionous use of registration_code
