### 2018-10-25 oxinabox/DataDepsGenerators.jl#67 @oxinabox
No Github #69 only fixed one problem
