### 2018-10-25 oxinabox/DataDepsGenerators.jl#67 @oxinabox
@SebastinSanty  I think we should just remove `DataOneV1()`
we have like 4 other APIs to get the same information.
most of the information in `DataOneV1()` is actually using `dcterms` etc in the DataDryad HTML, rather than the actual `DataOneV1()`
And DataDryad has recently completely changed how they use `dcterms` breaking it utterly.
It could be fixed, but I am not sure the point since it is redundant for DataDryad, which is all that it works for.

I think later we should consider making something that robustly works with `dcterms`, and works across sites.
