### 2018-10-25 oxinabox/DataDepsGenerators.jl#67 @SebastinSanty
Yes, we can safely remove `DataOneV1()` specially because as far as we know this DataOne version is only used by DataDryad. Also, DataDryad based registration blocks are generated even without that as you said.
