+++ src/DataDepsGenerators.jl
@@ -7,7 +7,7 @@ using Dates
 using InteractiveUtils: subtypes
 
 export generate, citation_text, remove_cite_version
-export UCI, GitHub, DataDryad, DataOneV1, DataOneV2, CKAN, DataCite, Figshare, JSONLD_Web, JSONLD_DOI
+export UCI, GitHub, DataDryad, DataOneV2, CKAN, DataCite, Figshare, JSONLD_Web, JSONLD_DOI
 
 
 
@@ -38,7 +38,6 @@ include("misc_extractors.jl")
 include("APIs/UCI.jl")
 include("APIs/GitHub.jl")
 include("APIs/DataDryad.jl")
-include("APIs/DataOneV1.jl")
 include("APIs/DataOneV2/DataOneV2.jl")
 include("APIs/CKAN.jl")
 include("APIs/DataCite.jl")
