### 2018-10-31 oxinabox/DataDepsGenerators.jl#67 @oxinabox
2/3 fixed. (well 1 fixed, 1 deleted)
All that is left is DataDepsV2
re: https://github.com/JuliaWeb/HTTP.jl/issues/342

According to Twitter on that DataOne wants to talk about the problem on there slack so I'll go talk to them.
