### 2018-10-31 oxinabox/DataDepsGenerators.jl#67 @datadavev
DataONE services use `SSLVerifyClient optional` for clients to optionally authenticate using a client certificate which is one reason for TLS renegotiation occurring. RFC 5746 identifies the TLS renegotiation extension that addresses the vulnerability described in CVE-2009-3555. RFC 5746 is implemented in OpenSSL versions 0.9.8m and 1.0.0a. Our systems are currently at OpenSSL 1.0.1f, so my understanding is that the vulnerability is resolved and hence the guidance to avoid TLS renegotiation should probably be revised.

Edit: The issue seems to be around TLS 1.3 which will not support renegotiation. It is not clear yet how Apache will support optional SSLVerifyClient with TLS 1.3.
