### 2018-11-05 #921 @oxinabox
@arfon  I fixed up those spelling issues (though, I'll give it a 3rd check),
I was just waiting to fix up an issue @ninjin identified with one of our data providers,
https://github.com/oxinabox/DataDepsGenerators.jl/issues/67
Something in the DataONE API, or in the SSL client we use to access it changed in a breaking way in the last few months.
I am talking with the devs on both sides there.
