### 2018-11-05 #921 @ninjin
Fix “lavages” – [fun fact](https://en.wiktionary.org/wiki/lavage#English). Other than that is looks solid.
