### 2018-11-05 #921 @ninjin
Looks good to me, @arfon do I sign off on the final version by closing this issue?
