### 2018-11-05 #921 @arfon
> Looks good to me, @arfon do I sign off on the final version by closing this issue?

Thanks @ninjin - I'll take it from here.
