### 2018-11-06 #921 @arfon
@ninjin - many thanks for your review here ✨

@oxinabox - your paper is now accepted into JOSS :zap: :rocket: :boom:
